pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognirename"
version = "0.1.0"
description = "顔認識による写真自動リネームツール"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CogniRename Project", email = "project@cognirename.local"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["face-recognition", "photo-management", "rename", "ai"]
requires-python = ">=3.11"

dependencies = [
    "face-recognition>=1.3.0",
    "dlib>=19.24.0", 
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "streamlit>=1.28.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
cognirename = "cognirename.cli.commands:main"

[project.urls]
Homepage = "https://github.com/cognirename/cognirename"
Repository = "https://github.com/cognirename/cognirename.git"
Documentation = "https://cognirename.readthedocs.io"
Issues = "https://github.com/cognirename/cognirename/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cognirename = ["py.typed"]

# Black コードフォーマット設定
[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

# MyPy 型チェック設定
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

# Pytest テスト設定
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 